Refactor product-summary-buy-button to TypeScript
