+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Refactor `product-summary-column` to TypeScript.
 - Refactor `product-summary-name` to TypeScript.
 - Refactor `product-summary-sku-selector` to TypeScript.
+- Refactor `product-summary-buy-button` to TypeScript.
 
 ## [2.66.3] - 2020-12-16
 ### Fixed
