--- react/ProductSummaryBuyButton.js
@@ -1,3 +1,0 @@
-import ProductSummaryBuyButton from './components/ProductSummaryBuyButton/ProductSummaryBuyButton'
-
-export default ProductSummaryBuyButton
